+++ CHANGES.md
@@ -5,6 +5,8 @@ unreleased
 0.28.0 (05/10/2022)
 -------------------
 
+- Add `Ppxlib.Mangle` with name mangling utilities from ppx_deriving (#<PR_NUMBER>, @sim642)
+
 - Make `esequence` right-associative. (#366, @ceastlund)
 
 - Deprecate unused attributes in `Deriving.Generator` (#368, @sim642)